Fix zero-length slice field with omitempty
There were two problems here, the combination of which caused incorrect
output.  The first is that we should really only check omitempty on a
struct field; a slice alone should ignore it.  For example
`repr.String([]int{})` should return `[]int{}`, not `""` (and likewise
for `[]int(nil)`).  The second is that the check in the slice code was
checking `len(s) == 0`, whereas the check in the struct-field code was
checking `s == nil`.  Either would be reasonable, but having both means
they need to match!  This generated output like
`repr.intSliceStruct{f: }` which is a syntax error.

Anyway, to fix this I just removed the check in slice.  We still omit
nil correctly, we just handle it at the point of the struct field.