feat: OmitZero and OmitEmpty (#30)
This splits OmitEmpty into OmitEmpty and OmitZero. OmitEmpty will omit
anything that is zero as well as a slice and map of length 0.

OmitZero only omits the zero type, either by checking the return value
of the IsZero method or if that's not implemented by
reflect.Value.IsZero. This by default retains the empty map and slice,
since initialised but 0-length maps and slices are not zero.